# REST APIs Design/openapi.yaml
openapi: 3.0.1
info:
  title: Film Manager
  description: REST APIs for the interaction with the Film Manager service.
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'https://todo.swagger.io/v2'
  - url: 'http://todo.swagger.io/v2'
...
paths:
...
  '/api/films/public': 
    get:
      tags:
        - films/public
      summary: Retrieve the public films
      description: >-
        The public films (i.e., the films that are visible for all the users of the service) are retrieved. This operation does not require authentication. A pagination mechanism is implemented to limit the size of messages.
      operationId: getPublicFilms
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: 'The id of the requested page (if absent, the first page is returned)'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  films:
                    type: array
                    items:
                      $ref: '#/components/schemas/Film'
                  next:
                    type: string
                    format: uri
              examples:
                Example:
                  value:
                    totalPages: 3
                    currentPage: 1
                    totalItems: 6
                    films:
                      - id: 1
                        title: "Your Name"
                        owner: 1
                        private: false
                        watchDate: "2021-10-03"
                        rating: 9
                        favorite: true
                        self: "/api/films/public/1"
                      - id: 2
                        title: "Heaven's Feel"
                        owner: 2
                        private: false
                        self: "/api/films/public/2"
                    next": "/api/films/public?pageNo=2"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
...
  '/api/films/public/{filmId}/reviews':
    post:
      tags:
        - 'films/public/{filmId}/reviews'
      security:
        - cookieAuth: []
      summary: Issue film review to some users
      description: The film with ID filmId is assigned to one or more users for review and the corresponding reviews are created. The users are specified in the review representations in the request body. This operation can only be performed by the owner.
      operationId: issueFilmReview
      parameters:
        - name: filmId
          in: path
          description: ID of the film
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: the new film reviews, including the users to whom they are issued and the filmId for which the reviews are being issued. 
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Review'
            examples:
              Example 1:
                value:
                  - filmId: 3
                    reviewerIds: [4,5] 
              Example 2:
                value:
                  - filmId: 3
                    reviewerIds: [4,5]   
                  - filmId: 3
                    reviewerIds: [6]    
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/Review'
              examples:
                Example 1:
                  value:
                    - filmId: 3
                      reviewId: 13
                      reviewerIds: [4,5]
                      completed: false
                      self: "/api/films/public/3/reviews/13"
                Example 2:
                  value:
                    - filmId: 3
                      reviewId: 14
                      reviewerIds: [3]
                      completed: false
                      self: "/api/films/public/3/reviews/14"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                invalid parameters:
                  value:
                    errors:
                      - param: Server
                        msg: Invalid filmId, it must be a number.
                missing parameters:
                  value:
                    errors:
                      - param: Server
                        msg: filmId parameter missing
                duplicate parameters: 
                  value:
                    errors:
                      - param: Server
                        msg: There are duplicate values of reviewerIds in group reviews.
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Not authorized'
              examples:
                Not authorized:
                  value:
                    error: 'Not authorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                Example:
                  value:
                    errors:
                      - param: Server
                        msg: The user is not the owner of the film.
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                Example:
                  value:
                    errors:
                      - param: Server
                        msg: The film does not exist.
        '409':
          description: Conflict (e.g., Unknown reviewer id)
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                Example:
                  value:
                    errors:
                      - param: Server
                        msg: The user/s with ID/S reviewerId/s do/es not exist.
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
    get:
      tags:
        - 'films/public/{filmId}/reviews'
      summary: Retrieve the list of all the reviews that have been issued/completed for a film 
      description: >-
        All the reviews that have been issued/completed for the film with ID filmId are retrieved. A pagination mechanism is implemented to limit the size of messages.
        This operation does not require authentication. 
      operationId: getFilmReviews
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose reviews must be retrieved
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: ID of the requested page (if absent, the first page is returned)'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                  type: array
                  items:
                    type: object
                    properties:
                      totalPages:
                        type: integer
                      currentPage:
                        type: integer
                      totalItems:
                        type: integer
                      reviews:
                        type: array
                        items:
                          $ref: '#/components/schemas/Review'
                      next:
                        type: string
                        format: uri
              examples:
                Example:
                  value:
                    totalPages: 1
                    currentPage: 1
                    totalItems: 2
                    reviews:
                      - filmId: 1
                        reviewId: 11
                        reviewerIds: [1]
                        completed: false
                        self: "/api/films/public/1/reviews/11"
                      - filmId: 1
                        reviewId: 12
                        reviewerIds: [1,2,3]
                        completed: false
                        self: "/api/films/public/1/reviews/12"
        '400': 
          description: Bad Request 
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                invalid parameters:
                  value:
                    errors:
                      - param: Server
                        msg: Invalid filmId, it must be a number.
                missing parameters:
                  value:
                    errors:
                      - param: Server
                        msg: filmId parameter missing
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
              examples:
                Example:
                  value:
                    errors:
                      - param: Server
                        msg: The page does not exist.
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        param:
                          type: string
                        msg:
                          type: string
...
components:
  schemas:
...
    Film:
      description: schema for Film data structures
      type: object
      additionalProperties: false
      properties:
        $schema:
          type: string
          description: path to the JSON Schema to use for the validation of the JSON data structure
            describing the film
        id:
          type: integer
          description: unique identifier for the film data structure
        title:
          type: string
          description: textual title of the film
        owner:
          type: integer
          description: unique identifier of the the film data structure owner
        private:
          type: boolean
          description: true if only the owner of the film data structure can access it, false otherwise
          default: 'true'
        watchDate:
          type: string
          format: date
          description: the date when the film was watched by the owner
        rating:
          type: integer
          description: the rating the owner has given to the film
          minimum: 1
          maximum: 10
        favorite:
          type: boolean
          description: 'true if the film is among the favourite
            ones of the owner, false otherwise '
          default: 'false'
        self: 
          type: string
          format: uri
      required:
      - title
      - private
    Review:
      description: schema for Review data structures
      type: object
      additionalProperties: false
      properties:
        $schema:
          type: string
          description: path to the JSON Schema to use for the validation of the JSON data structure
            describing the Review
        filmId:
          type: integer
          description: unique identifier of the film for which a review invitation has been
            issued
        reviewId:
          type: integer
          description: unique identifier for the review
        reviewerIds:
          type: array
          minItems: 1
          items:
              type: integer 
          description: unique identifiers of the the user or users who have received this review assignment
        completed:
          type: boolean
          description: true if the review has been completed, false
            otherwise
          default: 'false'
        reviewDate:
          type: string
          format: date
          description: the date when the review has been completed by the invited user
        rating:
          type: integer
          description: the rating the reviewer has given to the film
          minimum: 1
          maximum: 10
        review:
          type: string
          description: a textual description of the review
          maxLength: 1000
        self: 
          type: string
          format: uri
...
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid